requirements.txt
gradio==3.1.4
langchain-groq== 0.3.1
sentence-transformers==2.2.2
rank-bm25==0.2.2
pdfplumber==0.11.5
langchain==0.3.21
langchain-community==0.3.20
faiss-cpu==1.8.0.post1
dotenv
numpy>=1.26.2
huggingface-hub==0.25.2
pypdf==5.4.0